refactor(sqlite): retain boundary implementation owners
