+++ crates/atm-storage-rusqlite/src/lib.rs
@@ -31,10 +31,11 @@ mod search_store;
 mod shared_db;
 mod shared_db_diagnostics;
 mod shared_db_reader_lanes;
-mod storage_types;
 mod team_roster_schema;
 mod template_catalog_schema;
 mod template_catalog_store;
+#[cfg(any(test, feature = "test-support"))]
+mod test_support;
 mod writer;
 
 pub use reader_pool::{ReaderLaneMetricsSnapshot, ReaderLanesMetricsSnapshot};
@@ -75,11 +76,11 @@ use search_store::{async_search_store, search_store};
 use shared_db::{SharedDb, deserialize_json};
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
-use storage_types::{
-    SqliteMessageStore, SqliteNudgeTemplateOverrideStore, SqlitePeerConfigStore,
-    SqlitePendingNudgeStore, SqliteRosterStore, StoredMailMessageState,
-};
 use template_catalog_store::template_catalog_store;
+#[cfg(any(test, feature = "test-support"))]
+pub use test_support::{
+    TemplateAdmissionMessage, TemplateAdmissionSnapshot, inspect_template_admission_for_test,
+};
 pub use writer::DiagnosticTimelinePersistenceStats;
 
 #[cfg(any(test, feature = "test-support"))]
@@ -99,33 +100,6 @@ impl Drop for SqliteWriterLockGuard {
 #[cfg(any(test, feature = "test-support"))]
 pub struct TestOnlySqliteWriterLockGuard {
     _guard: SqliteWriterLockGuard,
-}
-
-/// Test-only projection of template-admission state.
-///
-/// Keeping this probe in SQLite test support lets the replacement HTTP runtime
-/// prove durable rows without importing SQLite or opening a database itself.
-#[doc(hidden)]
-#[cfg(any(test, feature = "test-support"))]
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct TemplateAdmissionSnapshot {
-    pub template_count: usize,
-    pub decomposed_count: usize,
-    pub messages: Vec<TemplateAdmissionMessage>,
-}
-
-/// One durable message projection returned by [`TemplateAdmissionSnapshot`].
-#[doc(hidden)]
-#[cfg(any(test, feature = "test-support"))]
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct TemplateAdmissionMessage {
-    pub message_key: String,
-    pub template_sha: Option<String>,
-    pub vars_json: Option<String>,
-    pub category: Option<String>,
-    pub content_format: Option<String>,
-    pub tags_json: String,
-    pub message_text: Option<String>,
 }
 
 #[cfg(any(test, feature = "test-support"))]
@@ -149,73 +123,6 @@ pub fn hold_sqlite_writer_lock_for_test(
     path: impl AsRef<Path>,
 ) -> Result<TestOnlySqliteWriterLockGuard, AtmError> {
     hold_sqlite_writer_lock(path).map(|guard| TestOnlySqliteWriterLockGuard { _guard: guard })
-}
-
-/// Reads only durable template-admission projections for black-box tests.
-///
-/// Production code must use the sealed storage contracts. This helper is
-/// test-support-only so the Tokio HTTP runtime never owns a database handle.
-#[doc(hidden)]
-#[cfg(any(test, feature = "test-support"))]
-pub fn inspect_template_admission_for_test(
-    path: impl AsRef<Path>,
-    message_keys: &[String],
-) -> Result<TemplateAdmissionSnapshot, AtmError> {
-    let connection = Connection::open(path.as_ref()).map_err(|error| {
-        AtmError::mailbox_read(format!(
-            "failed to inspect template-admission fixture: {error}"
-        ))
-    })?;
-    let (template_count, decomposed_count): (i64, i64) = connection
-        .query_row(
-            "SELECT
-                (SELECT COUNT(*) FROM message_templates),
-                (SELECT COUNT(*) FROM decomposed_messages)",
-            [],
-            |row| Ok((row.get(0)?, row.get(1)?)),
-        )
-        .map_err(|error| {
-            AtmError::mailbox_read(format!(
-                "failed to count template-admission fixture rows: {error}"
-            ))
-        })?;
-    let messages = message_keys
-        .iter()
-        .map(|message_key| {
-            connection
-                .query_row(
-                    "SELECT message_key, template_sha, vars_json, category, content_format,
-                            tags_json, message_text
-                     FROM mail_messages WHERE message_key = ?1",
-                    params![message_key],
-                    |row| {
-                        Ok(TemplateAdmissionMessage {
-                            message_key: row.get(0)?,
-                            template_sha: row.get(1)?,
-                            vars_json: row.get(2)?,
-                            category: row.get(3)?,
-                            content_format: row.get(4)?,
-                            tags_json: row.get(5)?,
-                            message_text: row.get(6)?,
-                        })
-                    },
-                )
-                .map_err(|error| {
-                    AtmError::mailbox_read(format!(
-                        "failed to inspect template-admission message '{message_key}': {error}"
-                    ))
-                })
-        })
-        .collect::<Result<Vec<_>, _>>()?;
-    Ok(TemplateAdmissionSnapshot {
-        template_count: usize::try_from(template_count).map_err(|_| {
-            AtmError::mailbox_read("template-admission fixture count exceeds usize range")
-        })?,
-        decomposed_count: usize::try_from(decomposed_count).map_err(|_| {
-            AtmError::mailbox_read("template-admission fixture count exceeds usize range")
-        })?,
-        messages,
-    })
 }
 
 /// Installs a test-only SQLite trigger that deterministically rejects mailbox
@@ -246,6 +153,39 @@ pub fn install_message_write_failure_for_test(path: impl AsRef<Path>) -> Result<
 
 #[cfg(test)]
 pub(crate) use mailbox_metadata_types::SqliteMailboxMetadataRow;
+
+#[derive(Debug)]
+struct SqliteMessageStore {
+    db: Arc<SharedDb>,
+}
+
+#[derive(Debug)]
+struct SqliteRosterStore {
+    db: Arc<SharedDb>,
+}
+
+#[derive(Debug)]
+struct SqliteNudgeTemplateOverrideStore {
+    db: Arc<SharedDb>,
+}
+
+#[derive(Debug)]
+struct SqlitePendingNudgeStore {
+    db: Arc<SharedDb>,
+}
+
+#[derive(Debug)]
+struct SqlitePeerConfigStore {
+    db: Arc<SharedDb>,
+}
+
+#[derive(Debug, Clone)]
+struct StoredMailMessageState {
+    read: bool,
+    pending_ack_at: Option<IsoTimestamp>,
+    acknowledged_at: Option<IsoTimestamp>,
+    expires_at: Option<IsoTimestamp>,
+}
 
 impl SqliteMessageStore {
     fn new(db: Arc<SharedDb>) -> Self {
